# swagger_server/models/user_public.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.role import Role  # noqa: F401,E501
from swagger_server import util


class UserPublic(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, username: str=None, role: Role=None, avatar_url: str=None, bio: str=None):  # noqa: E501
        """UserPublic - a model defined in Swagger

        :param id: The id of this UserPublic.  # noqa: E501
        :type id: str
        :param username: The username of this UserPublic.  # noqa: E501
        :type username: str
        :param role: The role of this UserPublic.  # noqa: E501
        :type role: Role
        :param avatar_url: The avatar_url of this UserPublic.  # noqa: E501
        :type avatar_url: str
        :param bio: The bio of this UserPublic.  # noqa: E501
        :type bio: str
        """
        self.swagger_types = {
            'id': str,
            'username': str,
            'role': Role,
            'avatar_url': str,
            'bio': str
        }

        self.attribute_map = {
            'id': 'id',
            'username': 'username',
            'role': 'role',
            'avatar_url': 'avatarUrl',
            'bio': 'bio'
        }
        self._id = id
        self._username = username
        self._role = role
        self._avatar_url = avatar_url
        self._bio = bio

    @classmethod
    def from_dict(cls, dikt) -> 'UserPublic':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserPublic of this UserPublic.  # noqa: E501
        :rtype: UserPublic
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this UserPublic.


        :return: The id of this UserPublic.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this UserPublic.


        :param id: The id of this UserPublic.
        :type id: str
        """

        self._id = id

    @property
    def username(self) -> str:
        """Gets the username of this UserPublic.


        :return: The username of this UserPublic.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username: str):
        """Sets the username of this UserPublic.


        :param username: The username of this UserPublic.
        :type username: str
        """

        self._username = username

    @property
    def role(self) -> Role:
        """Gets the role of this UserPublic.


        :return: The role of this UserPublic.
        :rtype: Role
        """
        return self._role

    @role.setter
    def role(self, role: Role):
        """Sets the role of this UserPublic.


        :param role: The role of this UserPublic.
        :type role: Role
        """

        self._role = role

    @property
    def avatar_url(self) -> str:
        """Gets the avatar_url of this UserPublic.


        :return: The avatar_url of this UserPublic.
        :rtype: str
        """
        return self._avatar_url

    @avatar_url.setter
    def avatar_url(self, avatar_url: str):
        """Sets the avatar_url of this UserPublic.


        :param avatar_url: The avatar_url of this UserPublic.
        :type avatar_url: str
        """

        self._avatar_url = avatar_url

    @property
    def bio(self) -> str:
        """Gets the bio of this UserPublic.


        :return: The bio of this UserPublic.
        :rtype: str
        """
        return self._bio

    @bio.setter
    def bio(self, bio: str):
        """Sets the bio of this UserPublic.


        :param bio: The bio of this UserPublic.
        :type bio: str
        """

        self._bio = bio