# swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Usuarios y Autorización
  description: ""
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/universidadpolitcnic-44f/Usuarios_y_autorizacion/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.test/v1
...
paths:
...
  /auth/register:
    post:
      tags:
      - Auth
      summary: Alta de cuenta
      operationId: auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Datos inválidos
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/login:
    post:
      tags:
      - Auth
      summary: Inicia sesión
      operationId: auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciales inválidas o MFA requerida
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Renueva access token con refresh token
      operationId: auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_refresh_body"
        required: true
      responses:
        "200":
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Refresh token inválido
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Cierra sesión actual
      operationId: auth_logout_post
      responses:
        "204":
          description: Sesión cerrada
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/verify-email/request:
    post:
      tags:
      - Auth
      summary: Envía email de verificación
      operationId: auth_verify_email_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/verifyemail_request_body"
        required: true
      responses:
        "204":
          description: Enviado si el email existe
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/verify-email/confirm:
    post:
      tags:
      - Auth
      summary: Confirma verificación de email
      operationId: auth_verify_email_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/verifyemail_confirm_body"
        required: true
      responses:
        "204":
          description: Email verificado
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/password/reset/request:
    post:
      tags:
      - Auth
      summary: Solicita reset de contraseña
      operationId: auth_password_reset_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reset_request_body"
        required: true
      responses:
        "204":
          description: Enviado si el email existe
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/password/reset/confirm:
    post:
      tags:
      - Auth
      summary: Confirma nuevo password
      operationId: auth_password_reset_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reset_confirm_body"
        required: true
      responses:
        "204":
          description: Contraseña actualizada
      x-openapi-router-controller: swagger_server.controllers.auth_controller
...
components:
  schemas:
    Role:
      type: string
      enum:
      - listener
      - artist
      - admin
    RegisterRequest:
      required:
      - email
      - password
      - role
      - username
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          minLength: 8
          type: string
        username:
          maxLength: 32
          minLength: 3
          type: string
        role:
          $ref: "#/components/schemas/Role"
    LoginRequest:
      required:
      - emailOrUsername
      - password
      type: object
      properties:
        emailOrUsername:
          type: string
        password:
          type: string
        mfaCode:
          type: string
          description: Código TOTP si MFA está activado
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      example:
        accessToken: accessToken
        refreshToken: refreshToken
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserPrivate"
        tokens:
          $ref: "#/components/schemas/TokenPair"
      example:
        tokens:
          accessToken: accessToken
          refreshToken: refreshToken
        user: ""
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        avatarUrl:
          type: string
          format: uri
        bio:
          type: string
      example:
        role: listener
        avatarUrl: http://example.com/aeiou
        bio: bio
        id: id
        username: username
    UserPrivate:
      allOf:
      - $ref: "#/components/schemas/UserPublic"
      - type: object
        properties:
          email:
            type: string
            format: email
          emailVerified:
            type: boolean
...
    auth_refresh_body:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    verifyemail_request_body:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          format: email
    verifyemail_confirm_body:
      required:
      - token
      type: object
      properties:
        token:
          type: string
    reset_request_body:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          format: email
    reset_confirm_body:
      required:
      - newPassword
      - token
      type: object
      properties:
        token:
          type: string
        newPassword:
          minLength: 8
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth